6-size.c

+++ 0x00-hello_world/6-size.c
@@ -13,11 +13,11 @@ int main(void);
 	long long int d;
 	float e;
 
-	printf("Size of a char: %ld bytes\n", sizeof(a));
-	printf("Size of an int: %ld bytes\n", sizeof(b));
-	printf("Size of a long int: %ld bytes\n", sizeof(c));
-	printf("Size of a long long int: %ld bytes\n", sizeof(d));
-	printf("Size of a float: %ld bytes\n", sizeof(e));
+	printf("Size of a char: %lu bytes\n", (unsigned long)sizeof(a));
+	printf("Size of an int: %lu bytes\n", (unsigned long)sizeof(b));
+	printf("Size of a long int: %lu bytes\n", (unsigned long)sizeof(c));
+	printf("Size of a long long int: %lu bytes\n", (unsigned long)sizeof(d));
+	printf("Size of a float: %lu bytes\n", (unsigned long)sizeof(e));
 
 	return (0);
 }
